+++ account_financial_report/report/aged_partner_balance_xlsx.py
@@ -10,8 +10,9 @@ class AgedPartnerBalanceXslx(models.AbstractModel):
     _name = 'report.a_f_r.report_aged_partner_balance_xlsx'
     _inherit = 'report.account_financial_report.abstract_report_xlsx'
 
-    def _get_report_name(self):
-        return _('Aged Partner Balance')
+    def _get_report_name(self, report):
+        report_name = _('Aged Partner Balance')
+        return self._get_report_complete_name(report, report_name)
 
     def _get_report_columns(self, report):
         if not report.show_move_line_details: