+++ account_financial_report/report/journal_ledger_xlsx.py
@@ -10,8 +10,9 @@ class JournalLedgerXslx(models.AbstractModel):
     _name = 'report.a_f_r.report_journal_ledger_xlsx'
     _inherit = 'report.account_financial_report.abstract_report_xlsx'
 
-    def _get_report_name(self):
-        return _('Journal Ledger')
+    def _get_report_name(self, report):
+        report_name = _('Journal Ledger')
+        return self._get_report_complete_name(report, report_name)
 
     def _get_report_columns(self, report):
         columns = [