+++ account_financial_report/report/vat_report_xlsx.py
@@ -8,8 +8,9 @@ class VATReportXslx(models.AbstractModel):
     _name = 'report.a_f_r.report_vat_report_xlsx'
     _inherit = 'report.account_financial_report.abstract_report_xlsx'
 
-    def _get_report_name(self):
-        return _('VAT Report')
+    def _get_report_name(self, report):
+        report_name = _('VAT Report')
+        return self._get_report_complete_name(report, report_name)
 
     def _get_report_columns(self, report):
         return {