+++ account_financial_report/readme/HISTORY.rst
@@ -1,3 +1,16 @@
+11.0.2.5.0 (2019-04-26)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* In the Trial Balance you have an option to hide parent hierarchy levels
+
+11.0.2.4.1 (2019-01-08)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* Handle better multicompany behaviour
+* Improve how title appears in the reports
+* Improve performance in General Ledger
+
+
 11.0.2.3.1 (2018-11-29)
 ~~~~~~~~~~~~~~~~~~~~~~~